+++ src/Cli/func/StaticResources/StaticResources.cs
@@ -48,6 +48,8 @@ namespace Azure.Functions.Cli
 
         public static Task<string> DockerfilePython313 => GetValue("Dockerfile.python3.13");
 
+        public static Task<string> DockerfilePython313BuildEnv => GetValue("Dockerfile.python3.13-buildenv");
+
         public static Task<string> DockerfilePowershell7 => GetValue("Dockerfile.powershell7");
 
         public static Task<string> DockerfilePowershell72 => GetValue("Dockerfile.powershell7.2");