+++ release_notes.md
@@ -8,3 +8,4 @@
 #### Changes
 
 - <entry>
+- Add Dockerfile for python 3.13 local build environment (#4611)